Control.Push

Documentation

class (Functor f, Functor g) => Part f g where

Minimal complete definition

part

Methods

part :: f (g a) -> f (f (g a))

class (Functor f, Functor g) => Dart f g where

Minimal complete definition

dart

Methods

dart :: f (f (g a)) -> f (g (f a))

class (Part f g, Dart f g) => Push f g where

Methods

push :: f (g a) -> f (g (f a))

class (Functor f, Functor g) => Draw f g where

Minimal complete definition

draw

Methods

draw :: f (g (f a)) -> f (f (g a))

class (Functor f, Functor g) => Bond f g where

Minimal complete definition

bond

Methods

bond :: f (f (g a)) -> f (g a)

class (Draw f g, Bond f g) => Pull f g where

Methods

pull :: f (g (f a)) -> f (g a)

newtype GPushedPlus f g h a

Constructors

GPushedPlus

Fields

runGPushedPlus :: (GPushedPlus1 g h f :+: GPushedPlus1 h g f) a

newtype GPushedPlus1 g h f a

Constructors

GPushedPlus1

Fields

runGPushedPlus1 :: ((f :.: (g :+: h)) :+: (Maybe :.: g)) a

newtype NonEmptyT f a

Constructors

NonEmptyT

Fields

runNonEmptyT :: (a, f a)

data Pair a

Constructors

Pair

Fields

pairFst :: !a

pairSnd :: !a

notes :: ()

Push one Functor through another.

The goal-function is push, which pushes a "layer" (part) of the outer Functor (f) "over" or "through" (pver) the inner Functor (g).

Some experiements and scratch-work proofs demonstrate that there are plenty of "exotic" cases, where ignoring part and pver cause one's intuition about pushing one Functor through another to break down.

For example, Either has multiple, simple implementations:

Requiring Monad, even though Applicative would be enough for this implementation:

  fmap (fmap return) :: (Functor f, Functor g, Monad m) => f         (g b) -> f         (g (m         b))
  fmap (fmap return) :: (Functor f, Functor g)          => f         (g b) -> f         (g (Either a' b))
  fmap (fmap return) ::             Functor g           => Either a' (g b) -> Either a' (g (Either a' b)) -- Note that t' does not have to be equivalent to t
  push               ::             Functor g           => Either a  (g b) -> Either a  (g (Either a' b))
  push               ::             Functor g           => Either a  (g b) -> Either a  (g (Either a  b))

Specializing (probably "too early") to Right since return seemed too abstract:

  fmap (fmap Right) :: (Functor f, Functor g) => f         (g b) -> f         (g (Either a b))
  fmap (fmap Right) ::             Functor g  => Either a' (g b) -> Either a' (g (Either a b))
  fmap (fmap Right) ::             Functor g  => Either a  (g b) -> Either a  (g (Either a b))
  push              ::             Functor g  => Either a  (g b) -> Either a  (g (Either a b))

Trying out specializing the left fmap first instead, i.e. outermost specialization (right to left), instead of the innermost specialization (left to right) found in 1. and 2.).

  either Left (return . fmap return) :: (Functor g, Monad m) => Either a' (g a) -> Either a' (g (m        b))
  either Left (Right  . fmap return) :: (Functor g, Monad m) => Either a' (g a) -> Either a' (g (m        b))
  either Left (return . fmap return) ::  Functor g           => Either a' (g b) -> Either a' (g (Either a b))
  either Left (return . fmap Right ) ::  Functor g           => Either a' (g b) -> Either a' (g (Either a b))
  either Left (Right  . fmap Right ) ::  Functor g           => Either a' (g b) -> Either a' (g (Either a b))
  either Left (Right  . fmap Right ) ::  Functor g           => Either a  (g b) -> Either a  (g (Either a b))
  push                               ::  Functor g           => Either a  (g b) -> Either a  (g (Either a b))

Misc. notes:

  t a
  substructure (t a)

  Ahh! For a (substructure -> Bool) partition into (True substructures, False substructures), this is just a push-through!

  - What about for (substructure -> Int)?
  - Or (substructure -> poset)?
  - Or (substructure -> set)?

  For a continuous, linear functor, (substructure -> [0,1::Real]) gives us a weighted filter on the functor.

  In theory, this could be used to factor sounds:
  - The continuous linear functor is a representation of the song's wave and the partition function is a recognizer
   (it outputs the percent certainty that a wave clip is in a given class).

  - Then, one can pull out (weighted) the matching and non-matching parts.
    * By reducing the intensity of subsections in proportion, inverse proportion resp.
      to the liklihood, we could derive an approximation of the removal, matching parts resp. of the matched sound(s).

Orphan instances

(Applicative f, Applicative g) => Applicative ((:+:) * f g)

Methods

pure :: a -> (* :+: f) g a

(<*>) :: (* :+: f) g (a -> b) -> (* :+: f) g a -> (* :+: f) g b

liftA2 :: (a -> b -> c) -> (* :+: f) g a -> (* :+: f) g b -> (* :+: f) g c

(*>) :: (* :+: f) g a -> (* :+: f) g b -> (* :+: f) g b

(<*) :: (* :+: f) g a -> (* :+: f) g b -> (* :+: f) g a