Scratch

Documentation

s1 :: ()

 push :: f (Either a b) -> f (Either (f a) (f b))

 we need to be able to:
   f a -> f (StronglyConnectedComponent f a)

 which is two steps:
   isolate:
     we need to isolate the SCC
   constract:
     we need to replace the SCC with a single node containing the SCC

 what the end-functor construction does is replace the contraction with an equivalent operation: redefinition of data boundaries
   with an added bonus: the boundaries are "free", and always result in an isomorphic type (proof?)

endFunctorDef :: ()

 An end-functor is a functor with the following properties:
   It has an end value, whose type is solely dependent upon f:
     type EndOf f :: *
     endOf :: f a -> EndOf f
     end :: EndOf f -> f a
     endOf . end == id
   Its end value may be replaced with a new one, as a result of a fold:
     Bifunctor (End f)
     type EndOf (End f e) = e
     foldEnd :: Fold a (EndOf f -> e) -> f a -> End f e a
     foldEndM :: Monad m => FoldM a (EndOf f -> e) -> f a -> m (End f e a)
   Its middle values may be split by an end, resulting in a duplicate-like operation:
     foldEnds :: (a -> Bool) -> f a -> FixEnd f a
     foldEndsM :: Monad m => (a -> m Bool) -> f a -> m (FixEnd f a)

     FoldM (Either (Left a)) b () -> FoldM (Either (Right b)) a () -> f (Either a b) -> EitherFlip (FixEnd2 f) a b

s2 :: ()

 what about:
   newtype Stream a = Stream { runStream :: (a, Stream a) }
 we can push it through Either:
   Stream (Either a b) -> NonEmptyEnd (NonEmptyEnd .. b) a

 newtype EitherFlip f a b = EitherFlip { getEitherFlip :: Either (f b a) (f a b) } deriving (Eq, Ord, Show, Functor)

 newtype FixEnd f a = FixEnd { runFixEnd :: End f (FixEnd f a) a }

 newtype FixEnd2 f a b = FixEnd2 { runFixEnd2 :: End f (FixEnd2 f b a) b }

 newtype SomeIso a = SomeIso { runSomeIso :: (a -> b, b, b -> a) }